# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MDIA API",
        "description": "API documentation for Mbaise Diaspora in the Americas application",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/payments/submit": {
            "post": {
                "tags": [
                    "Member Payments"
                ],
                "summary": "Submit a payment with evidence",
                "description": "Submit a payment on behalf of authenticated member with evidence files. Status is automatically set to pending until verified.",
                "operationId": "ea6c1f7047d67d111495d833231f4262",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "payment_type_uuid",
                                    "year",
                                    "payment_date",
                                    "evidence_files"
                                ],
                                "properties": {
                                    "payment_type_uuid": {
                                        "type": "string",
                                        "format": "uuid",
                                        "example": "9d4e8b32-3c7a-4f9e-8b1a-2d3e4f5a6b7c"
                                    },
                                    "year": {
                                        "type": "integer",
                                        "example": 2024
                                    },
                                    "payment_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-01-15"
                                    },
                                    "notes": {
                                        "type": "string",
                                        "example": "Payment via bank transfer",
                                        "nullable": true
                                    },
                                    "evidence_files": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "example": "/uploads/evidence/receipt_123.jpg"
                                        },
                                        "minItems": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Payment submitted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Payment submitted successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "payment": {
                                                    "properties": {
                                                        "uuid": {
                                                            "type": "string",
                                                            "format": "uuid"
                                                        },
                                                        "user_id": {
                                                            "type": "integer"
                                                        },
                                                        "payment_type_id": {
                                                            "type": "integer"
                                                        },
                                                        "amount": {
                                                            "type": "number",
                                                            "format": "float"
                                                        },
                                                        "year": {
                                                            "type": "integer"
                                                        },
                                                        "payment_date": {
                                                            "type": "string",
                                                            "format": "date"
                                                        },
                                                        "notes": {
                                                            "type": "string",
                                                            "nullable": true
                                                        },
                                                        "status": {
                                                            "type": "string",
                                                            "example": "pending"
                                                        },
                                                        "verified_by": {
                                                            "type": "integer",
                                                            "nullable": true
                                                        },
                                                        "verified_at": {
                                                            "type": "string",
                                                            "format": "date-time",
                                                            "nullable": true
                                                        },
                                                        "created_at": {
                                                            "type": "string",
                                                            "format": "date-time"
                                                        },
                                                        "updated_at": {
                                                            "type": "string",
                                                            "format": "date-time"
                                                        },
                                                        "evidences": {
                                                            "type": "array",
                                                            "items": {
                                                                "properties": {
                                                                    "uuid": {
                                                                        "type": "string",
                                                                        "format": "uuid"
                                                                    },
                                                                    "payment_id": {
                                                                        "type": "integer"
                                                                    },
                                                                    "file_path": {
                                                                        "type": "string"
                                                                    },
                                                                    "created_at": {
                                                                        "type": "string",
                                                                        "format": "date-time"
                                                                    },
                                                                    "updated_at": {
                                                                        "type": "string",
                                                                        "format": "date-time"
                                                                    }
                                                                },
                                                                "type": "object"
                                                            }
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The payment type uuid field is required."
                                        },
                                        "errors": {
                                            "properties": {
                                                "payment_type_uuid": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The selected payment type uuid is invalid."
                                                    }
                                                },
                                                "evidence_files": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The evidence files field is required."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}